0.6.1 - breaking protocol amendment bug fix
technically a breaking change. now `BestMove` is an enum with a `Other` and a `Normal(NormalBestMove)` variant, where `NormalBestMove` is the old `BestMove`. see the `Other` variant's docs for why this was done. it boils down to the fact that engines dont follow protocol because the protocol is bad and they have to make their own `bestmove` messages when dealing with invalid positions. also added a test for this