renamed `Engine` fields + additional bounds on `traits::Message`
engine fields weren't very clear before. You would probably think `Engine.in` is the input to the engine, but nope. I just called it that because  we read from the engine, I guess that's `in`. But then it should be called a `Gui`.

additional trait bounds make it more convenient to use, really no downside because it's sealed